third_party/blink/web_tests/webmidi/open-close-expected.txt

CONSOLE WARNING: line 59: Web MIDI will ask a permission to use even if the sysex is not specified in the MIDIOptions since M77, around September 2019. See https://www.chromestatus.com/feature/5138066234671104 for more details.
Tests MIDIPort.open and MIDIPort.close.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PermissionsHelper.setPermission is defined.
PASS requestMIDIAccess() succeeded with access [object MIDIAccess].
Check state transition for close on closed state.
- check initial state.
PASS port.connection is "closed"
- check callback arguments.
PASS callback is called with port [object MIDIInput].
PASS callbackport.id is "MockInputID-0"
PASS callbackport.connection is "closed"
- check final state.
PASS port.connection is "closed"
Check state transition for open on closed state.
- check initial state.
PASS port.connection is "closed"
- check port handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "open"
- check access handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "open"
- check callback arguments.
PASS callback is called with port [object MIDIInput].
PASS callbackport.id is "MockInputID-0"
PASS callbackport.connection is "open"
- check final state.
PASS port.connection is "open"
Check state transition for open on open state.
- check initial state.
PASS port.connection is "open"
- check callback arguments.
PASS callback is called with port [object MIDIInput].
PASS callbackport.id is "MockInputID-0"
PASS callbackport.connection is "open"
- check final state.
PASS port.connection is "open"
Check state transition for close on open state.
- check initial state.
PASS port.connection is "open"
- check port handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "closed"
- check access handler.
PASS handler is called with port [object MIDIInput].
PASS eventport.id is "MockInputID-0"
PASS eventport.connection is "closed"
- check callback arguments.
PASS callback is called with port [object MIDIInput].
PASS callbackport.id is "MockInputID-0"
PASS callbackport.connection is "closed"
- check final state.
PASS port.connection is "closed"
PASS successfullyParsed is true

TEST COMPLETE
